refactor logs panel and diffs
